+++ lib/src/lints/avoid_unused_parameters/avoid_unused_parameters_rule.dart
@@ -1,6 +1,7 @@
 import 'package:analyzer/error/listener.dart';
 import 'package:custom_lint_builder/custom_lint_builder.dart';
 import 'package:solid_lints/src/lints/avoid_unused_parameters/visitors/avoid_unused_parameters_visitor.dart';
+import 'package:solid_lints/src/models/ignored_entities_model/ignored_entities_model.dart';
 import 'package:solid_lints/src/models/rule_config.dart';
 import 'package:solid_lints/src/models/solid_lint_rule.dart';
 
@@ -64,7 +65,7 @@ import 'package:solid_lints/src/models/solid_lint_rule.dart';
 /// };
 ///
 /// ```
-class AvoidUnusedParametersRule extends SolidLintRule {
+class AvoidUnusedParametersRule extends SolidLintRule<IgnoredEntitiesModel> {
   /// The [LintCode] of this lint rule that represents
   /// the error whether we use bad formatted double literals.
   static const String lintName = 'avoid_unused_parameters';
@@ -79,6 +80,7 @@ class AvoidUnusedParametersRule extends SolidLintRule {
     final rule = RuleConfig(
       configs: configs,
       name: lintName,
+      paramsParser: IgnoredEntitiesModel.fromJson,
       problemMessage: (_) => 'Parameter is unused.',
     );
 
@@ -92,7 +94,7 @@ class AvoidUnusedParametersRule extends SolidLintRule {
     CustomLintContext context,
   ) {
     context.registry.addCompilationUnit((node) {
-      final visitor = AvoidUnusedParametersVisitor();
+      final visitor = AvoidUnusedParametersVisitor(config.parameters);
       node.accept(visitor);
 
       for (final element in visitor.unusedParameters) {