+++ lib/src/lints/number_of_parameters/number_of_parameters_rule.dart
@@ -65,6 +65,11 @@ class NumberOfParametersRule
     CustomLintContext context,
   ) {
     context.registry.addDeclaration((node) {
+      if (config.parameters.ignoredEntitiesModel.isIgnoredClass(node) ||
+          config.parameters.ignoredEntitiesModel.isIgnoredMethod(node)) {
+        return;
+      }
+
       final parameters = switch (node) {
         (final MethodDeclaration node) =>
           node.parameters?.parameters.length ?? 0,