+++ CHANGELOG.md
@@ -8,6 +8,14 @@
 - add quick fix to avoid_final_with_getter (https://github.com/solid-software/solid_lints/pull/164)
 - Renamed `avoid_debug_print` to `avoid_debug_print_in_release`
 - The `avoid_debug_print_in_release` no longer reports a warning if the `debugPrint` call is wrapped in a `!kReleaseMode` check.
+- Consistent `exclude` configuration for the following rules:
+  - `avoid_returning_widgets`
+  - `avoid_unused_parameters`
+  - `cyclomatic_complexity`
+  - `function_lines_of_code`
+  - `no_empty_bloc`
+  - `number_of_parameters`
+
 
 ## 0.1.5